+++ changelog.d/20230412_100608_regis_palm.md
@@ -7,6 +7,8 @@
         - `APP` was previously a ContextTemplate, and is now a dictionary of contexts indexed by name. Developers who implemented this context should replace `Contexts.APP(...)` by `Contexts.app(...)`.
         - Removed the `ENV_PATCH` filter, which was for internal use only anyway.
         - The `PLUGIN_LOADED` ActionTemplate is now an Action which takes a single argument. (the plugin name)
-    - 💥[Refactor] We refactored the hooks API further by removing the static hook indexes and the hooks names. As a consequence:
-        - The syntactic sugar functions from the "filters" and "actions" modules were all removed: `get`, `add*`, `iterate*`, `apply*`, `do*`, etc.
+    - 💥[Refactor] We refactored the hooks API further by removing the static hook indexes and the hooks names. As a consequence, the syntactic sugar functions from the "filters" and "actions" modules were all removed: `get`, `add*`, `iterate*`, `apply*`, `do*`, etc.
     - 💥[Deprecation] The obsolete filters `COMMANDS_PRE_INIT` and `COMMANDS_INIT` have been removed. Plugin developers should instead use `CLI_DO_INIT_TASKS` (with suitable priorities).
+    - 💥[Feature] The "openedx" Docker image is no longer built with docker-compose in development on `tutor dev start`. This used to be the case to make sure that it was always up-to-date, but it introduced a discrepancy in how images were build (`docker compose build` vs `docker build`). As a consequence:
+        - The "openedx" Docker image in development can be built with `tutor images build openedx-dev`.
+        - The `tutor dev/local start --skip-build` option is removed. It is replaced by opt-in `--build`.
